feat: VideoMapper 생성 및 변환 로직 통합 #130
